Refactor types and headers
